refactor(groups): remove and consolidate excessive instances of ContactsRepository
