+++ CustomContacts/Code/UI/Main/Contacts/ContactSelectorView.swift
@@ -65,8 +65,6 @@ extension ContactSelectorView: Identifiable {
 extension ContactSelectorView {
 	// Could be `private` if not for Xcode warnings re: @Observable
 	@Observable final class ViewModel {
-		@ObservationIgnored @Dependency(\.contactsRepository) private var contactsRepository
-
 		private var contacts: [Contact] = []
 		var searchText = ""
 		private(set) var error: Error?
@@ -84,6 +82,7 @@ extension ContactSelectorView {
 
 		@MainActor func loadContacts(refresh: Bool = false) async {
 			do {
+				@Dependency(\.contactsRepository) var contactsRepository
 				contacts = try await contactsRepository.getAllContacts(refresh)
 			} catch {
 				self.error = error
