+++ CustomContacts/Code/UI/Main/Groups/GroupCardView.swift
@@ -7,26 +7,25 @@
 //
 
 import CustomContactsModels
-import Dependencies
 import SwiftUI
 
 struct GroupCardView: View {
 	@EnvironmentObject private var groupNavigation: GroupListNavigation
-	@Dependency(\.contactsRepository) private var contactsRepository
-	let group: ContactGroup
-	let onGroupTapped: () -> Void
-
 	@State private var isExpanded = false
+
+	private let group: ContactGroup
+	private let onGroupTapped: () -> Void
+
+	init(group: ContactGroup, onGroupTapped: @escaping () -> Void) {
+		self.group = group
+		self.onGroupTapped = onGroupTapped
+	}
 
 	var body: some View {
 		DisclosureGroup(
 			isExpanded: $isExpanded,
 			content: {
-				ForEach(
-					group.contactIDs
-						.compactMap { contactsRepository.getContact($0) }
-						.sorted()
-				) { contact in
+				ForEach(group.contacts().sorted()) { contact in
 					Button(
 						action: {
 							groupNavigation.path.append(.contactDetail(contact))
